remove parallel from_group_dataframe
