+++ darts/dataprocessing/transformers/fittable_data_transformer.py
@@ -10,7 +10,7 @@ import numpy as np
 
 from darts import TimeSeries
 from darts.logging import get_logger, raise_if, raise_if_not
-from darts.utils.parallel import _build_tqdm_iterator, _parallel_apply
+from darts.utils import _build_tqdm_iterator, _parallel_apply
 
 from .base_data_transformer import BaseDataTransformer
