+++ darts/utils/__init__.py
@@ -2,9 +2,9 @@
 Utils
 -----
 """
-# from .utils import (
-#     _build_tqdm_iterator,
-#     _parallel_apply,
-#     _with_sanity_checks,
-#     retain_period_common_to_all,
-# )
+from .utils import (
+    _build_tqdm_iterator,
+    _parallel_apply,
+    _with_sanity_checks,
+    retain_period_common_to_all,
+)
